First example of a deep learning system using R
This is an example from the excellent book Deep Learning with R, and shows the first steps in a deep learning network. The example is the well known NMIST data set. The final accuracy of the test data set is 0.9795. This will be improved in future statistical models of the same data set.